chore: BigInteger 타입을 Long으로 일괄 변경

+++ src/main/java/com/dayaeyak/waiting/domain/dto/request/WaitingCreateRequestDto.java
@@ -2,12 +2,11 @@ package com.dayaeyak.waiting.domain.dto.request;
 
 import lombok.Getter;
 
-import java.math.BigInteger;
 
 @Getter
 public class WaitingCreateRequestDto {
-    private BigInteger restaurantId;
-    private BigInteger datesId;
-    private BigInteger userId;
+    private Long restaurantId;
+    private Long datesId;
+    private Long userId;
     private Integer userCount;
 }
